Consolidated docker-compose configurations into one file with volume declarations in override file.
Since we use docker-compose only for configuration of single-use containers, not to launch
the app itself, there's no need to separate out the test and app containers into different files.
Consolidating them and moving volumes to an override file makes it cleaner and easier to
run tests against pure containers in pipelines and use volumes in development.