Rename test service to development and move to override docker-compose config.
The rename reflects the fact that this service may be used to run any npm scripts, not only tests.

Moving it to the override config means the main docker-compose file contains only instructions to
build the production image, reducing the chance of accidentally pushing dev/test images from the pipeline.