#ELM-699 Switch to Jest unit testing (#3)
* Reworded Tasks section of readme.

* Add script to run arbitrary commands in an API container instance.

For example, this enables npm commands to be run without installing npm directly onto a dev machine.

* Switch Lab to Jest for unit testing.

* Remove project-name and port from pipeline DockerCompose tasks.

* Enable arguments to be passed to docker-compose run commands via scripts.

* Debug file listing in pipeline built container image.

Tracing source of errors due to apparent lack of node_modules.

* Correct `containerCommand` input name for pipeline debug task.

* Properly wrap commands passed to build and test scripts and pipeline debug task.

* Escape quotes in pipeline debug command string.

* Removed echo statements from pipeline debug command.

* Split up pipeline debug into multiple stages.

* Consolidated docker-compose configurations into one file with volume declarations in override file.

Since we use docker-compose only for configuration of single-use containers, not to launch
the app itself, there's no need to separate out the test and app containers into different files.
Consolidating them and moving volumes to an override file makes it cleaner and easier to
run tests against pure containers in pipelines and use volumes in development.

* Remove debug stages from pipeline.

* Move Jest config into package.json.

Reordered package.json since it was looking a bit jumbled. The new order is as follows:
1. Descriptive elements (name, description, version, contributors, etc.)
2. Scripts
3. Dependencies
4. Configs for specific dependencies (Jest and StandardJS)

* Replaced pretest npm script with use of npm-run-all to run linting and unit tests in parallel.

This gives faster test runs and can be extended to run additional commands if required.
e.g. `npm-run-all test:*`

* Mount test-output/ folder from host into unit-test container for pipeline as well as dev.

* Add pipeline stage to publish Cobertura report and fix JUnit result publishing.

* Change watch:unit task to run Jest in watch mode.

* Use a variable for test-output folder in pipeline config.

* Set test task to run component tasks in parallel.

* Clean cache after npm install to minimise container image size.

* Add watch task to `bin/` scripts.

* Stop producing coverage reports in watch task to speed up feedback cycle.

* Rename test service to development and move to override docker-compose config.

The rename reflects the fact that this service may be used to run any npm scripts, not only tests.

Moving it to the override config means the main docker-compose file contains only instructions to
build the production image, reducing the chance of accidentally pushing dev/test images from the pipeline.

* Correct project name in watch task.

* Add override docker-compose file to test stages in pipeline.